Custom low battery warnings

+++ MusicNotch/Helpers/Defaults+Extension.swift
@@ -56,7 +56,7 @@ extension Defaults.Keys {
     
     static let batteryExtension = Key<Bool>("batteryExtension", default: true)
     static let lowPowerWarning = Key<Bool>("lowPowerWarning", default: true)
-    static let lowBatteryThreshold = Key<Int>("lowBatteryThreshold", default: 10)
+    static let lowBatteryThreshold = Key<Double>("lowBatteryThreshold", default: 10)
     static let lowPowerSound = Key<Bool>("lowPowerSound", default: true)
     static let pluggedInSound = Key<Bool>("pluggedInSound", default: true)
